requirements.txt
numpy==1.26.4
PyAutoGUI==0.9.54
python-dotenv==1.0.0
opencv-python==4.8.1.78
Pillow==10.1.0
